restructure seperating the timer types into differnt components
